requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118 
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
diffusers==0.27.2
accelerate==0.28.0
tensorflow==2.12.0
tensorboard==2.12.0
opencv-python==4.9.0.80
soundfile==0.12.1
transformers==4.39.2

gdown
requests
imageio[ffmpeg]

huggingface-hub==0.24.6

omegaconf
ffmpeg-python
gradio
spaces
moviepy

numpy
pillow
tqdm
matplotlib

insightface==0.7.3
